Migrate CLI to own module
