+++ nomenclature/__init__.py
@@ -5,7 +5,7 @@ from importlib.metadata import version
 
 from nomenclature.codes import CodeList  # noqa
 from nomenclature.core import DataStructureDefinition, create_yaml_from_xlsx  # noqa
-from nomenclature.testing import assert_valid_yaml  # noqa
+from nomenclature.cli import cli  # noqa
 from nomenclature.region_mapping_models import (  # noqa
     RegionProcessor,
     RegionAggregationMapping,
