+++ nomenclature/testing.py
@@ -2,26 +2,9 @@ import yaml
 import logging
 from pathlib import Path
 
-import click
-
 import nomenclature
 
 logger = logging.getLogger(__name__)
-
-cli = click.Group()
-
-
-@cli.command("validate-yaml")
-@click.argument("path", type=click.Path(exists=True, path_type=Path))
-def cli_valid_yaml(path: Path):
-    assert_valid_yaml(path)
-
-
-@cli.command("validate-project")
-@click.argument("path", type=click.Path(exists=True, path_type=Path))
-def cli_valid_project(path: Path):
-    assert_valid_yaml(path)
-    assert_valid_structure(path)
 
 
 def assert_valid_yaml(path: Path):
