+++ setup.cfg
@@ -29,10 +29,11 @@ tests =
     pytest
 docs =
     sphinx
+    sphinx-click
 
 [flake8]
 max-line-length = 88
 
 [options.entry_points]
 console_scripts =
-    nomenclature = nomenclature.testing:cli
+    nomenclature = nomenclature:cli
